QuantumBot/requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-groq==0.2.1
langchain-text-splitters==0.3.11
langsmith==0.1.147
streamlit==1.40.0
faiss-cpu==1.9.0.post1
pypdf==5.1.0
beautifulsoup4
python-dotenv==1.0.1
numpy==1.26.4
wikipedia
arxiv